Close gaps and cleanup tests, fix typos and comments, combine files.
